+++ source/native/symbulation_sgp.cc
@@ -5,7 +5,6 @@
 #include "../default_mode/Symbiont.h"
 #include "../sgp_mode/DiversityAnalysis.h"
 #include "../sgp_mode/ModularityAnalysis.h"
-#include "../sgp_mode/SGPDataNodes.h"
 #include "../sgp_mode/SGPWorld.h"
 #include "../sgp_mode/Scheduler.h"
 #include "../sgp_mode/SymbiontImpact.h"
@@ -14,7 +13,9 @@
 // Empirical doesn't support more than one translation unit, so any CC files are
 // included last. It still fixes include issues, but doesn't improve build time.
 #include "../default_mode/WorldSetup.cc"
+#include "../sgp_mode/SGPWorld.cc"
 #include "../sgp_mode/SGPWorldSetup.cc"
+#include "../sgp_mode/SGPWorldDataNodes.cc"
 #include "../sgp_mode/Tasks.cc"
 
 #include <fstream>
