+++ source/sgp_mode/SGPWorldSetup.cc
@@ -197,14 +197,6 @@ void SGPWorld::SetupSymbionts(unsigned long *total_syms) {
 
 }
 
-void SGPWorld::SendToGraveyard(emp::Ptr<Organism> org) {
-  if (org.DynamicCast<SGPSymbiont>()->GetCPU().state.in_progress_repro != -1) {
-    to_reproduce[org.DynamicCast<SGPSymbiont>()->GetCPU().state.in_progress_repro].second =
-      emp::WorldPosition::invalid_id;
-  }
-  SymWorld::SendToGraveyard(org);
-}
-
 int SGPWorld::GetNeighborHost(size_t id, emp::Ptr<Organism> symbiont) {
   // TODO - Why magic number here?
   // Attempt to find host that matches some tasks
